Added javadoc comments and removed redundant class
DirectionParcer had same functionality as DirectionJSONParcer, hence deleted.
Also attempted to move the functionality of ShowPlacesOnMapActivity, in order to work directly on the main activity. However, this causes the app to break in a way like the app closes on back button.
Currently those codes in main activity are commented until further fix.